+++ src/ch.ts
@@ -4,8 +4,9 @@ export {  isBoolean,
           isSymbol,
           isValidNumber } from './primitive';
 
-export {  isFunction,
-          isObject, 
+export {  isFunction } from './function';
+
+export {  isObject, 
           isNil } from './structural';
 
 export {  isArray } from './array';
@@ -24,6 +25,7 @@ export {  isJson,
           isRegex,
           isEmail,
           isIpAddress,
+          isJWT,
           isSlug,
           containsUpperCase,
           containsLowerCase,