+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 2.21.0 (Oct 18th 2023)
+
+- Add isJWT() method
+- Improve isArray() method
+- Update build tools
+- Code cleanup for lighter library
+- Add more unit tests
+
 # 2.20.0 (Oct 11th 2023)
 
 - Add typeCheck parameter to isTimestamp() and isValidTimestamp() methods
@@ -5,8 +13,8 @@
 
 # 2.19.2 (May 8th 2023)
 
-- Fix isNumber() & isTimestamp issue with array of length one
-- Add more tests with arrays
+- Fix isNumber() & isTimestamp() issue with array of length one
+- Add more unit tests with arrays
 
 # 2.19.1 (October 23th 2022)